Add files via upload
Disc A, with its center passing through the origin, is uniformly and symmetrically rotating within the x, y plane at a speed of V3 (1 kilo-revolution per second). At the beginning of the experiment, a force with a direction parallel to the Z axis and perpendicular to the x, y axes is applied directly above the Z axis to rotate the upper half of the disc around the x axis in the positive direction of the Z axis, while the lower half rotates in the negative direction of the Z axis. After rotating at a 45-degree angle, it stops (due to the gyroscopic effect, there is a tendency to return to the original state, requiring external force to fix it). Disc B (when the rotational speed is zero, a force with a direction parallel to the Z axis and perpendicular to the x, y axes is applied directly above the Z axis to first rotate the upper half of the disc around the x axis in the positive direction of the Z axis, while the lower half rotates in the negative direction of the Z axis. After rotating at a 45-degree angle, it stops, and then begins to accelerate and rotate to a speed of V3 and maintain it.) The observed and measured motion states of the two discs are exactly the same, but the gyroscopic effect determines and confirms that their actual motion states must be different。