Add files via upload
The question of why bicycles do not fall has puzzled scientists for two hundred years: why bicycles do not fall while riding has never been understood. When studying the AI algorithm of electric unicycles, it was found that both electric unicycles and manual unicycles follow the same physical theories in acceleration, deceleration, imbalance correction, and steering. The actual control process logic is also the same, except that the control subjects of electric unicycles are different. The control subjects of electric unicycles are sensors and microcomputers, while the control subjects of manual unicycles are cyclists. Based on the premise principle that the basic laws of physics in daily life behavior must be followed as objective laws, this study deeply analyzes and confirms the psychological and behavioral action processes of bicycle (human powered unicycle) players in the learning process using basic knowledge of physics mechanics and control theory.
In studying the behavior control process of cyclists during cycling, mature physics vector laws, control theory, systems science, and computer science are used to verify the behavior control process and related psychological activities of cyclists during cycling activities. Using basic knowledge of physics, it is confirmed that the essence of motion control in the abstract physics of cycling and satellite launch processes is the same, both of which are processes of object motion and attitude control. Objectively, when an object moves from the previous state of motion to the next state of motion, it must first break the balance of the previous state of motion. It is proposed that the result of cyclists learning how to ride a bicycle is not only to learn how to control the balance while riding a bicycle, but also to learn how to break the balance while riding a bicycle。